+++ tests/test_financial_calculations/test_bond_cash_flows_and_pricing.py
@@ -1,9 +1,8 @@
 import unittest
-from datetime import datetime, timedelta
 import numpy as np
+from datetime import datetime
 from financial_calculations.bond_cash_flows_and_pricing import (
-    create_spot_semi_bond_cash_flows,
-    discount_cash_flows
+    create_semi_bond_cash_flows
 )
 
 class TestCreateSpotSemiBondCashFlows(unittest.TestCase):
@@ -44,7 +43,7 @@ class TestCreateSpotSemiBondCashFlows(unittest.TestCase):
         
         expected_cash_flows = np.array([25000.0] * 9 + [1025000.0])
         
-        payment_dates, cash_flows = create_spot_semi_bond_cash_flows(
+        payment_dates, cash_flows = create_semi_bond_cash_flows(
             self.market_close_date, 
             self.balance, 
             self.coupon, 
@@ -52,7 +51,7 @@ class TestCreateSpotSemiBondCashFlows(unittest.TestCase):
         )
         
         # Check if the payment dates match
-        np.array_equal(payment_dates, expected_dates)
+        np.testing.assert_array_equal(payment_dates, expected_dates)
         
         # Check if the cash flows match
         np.testing.assert_array_equal(cash_flows, expected_cash_flows)
@@ -66,7 +65,7 @@ class TestCreateSpotSemiBondCashFlows(unittest.TestCase):
         to avoid potential end-of-month issues.
         """
         with self.assertRaises(ValueError):
-            create_spot_semi_bond_cash_flows("2024-08-29", self.balance, self.coupon, self.maturity_years)
+            create_semi_bond_cash_flows("2024-08-29", self.balance, self.coupon, self.maturity_years)
     
     def test_coupon_as_percentage(self):
         """
@@ -81,7 +80,7 @@ class TestCreateSpotSemiBondCashFlows(unittest.TestCase):
         expected_error_message = "Coupon should not be greater than 1 as it should be a decimal and not a percentage."
 
         with self.assertRaises(ValueError) as context:
-            create_spot_semi_bond_cash_flows(
+            create_semi_bond_cash_flows(
                 self.market_close_date, 
                 self.balance, 
                 coupon_as_percentage, 
@@ -104,21 +103,21 @@ class TestCreateSpotSemiBondCashFlows(unittest.TestCase):
         dt_market_close_date = datetime(2024, 8, 10)
         str_market_close_date = '2024-08-10'
 
-        np_payment_dates, np_cash_flows = create_spot_semi_bond_cash_flows(
+        np_payment_dates, np_cash_flows = create_semi_bond_cash_flows(
             np_market_close_date, 
             self.balance, 
             self.coupon, 
             self.maturity_years
         )
 
-        dt_payment_dates, dt_cash_flows = create_spot_semi_bond_cash_flows(
+        dt_payment_dates, dt_cash_flows = create_semi_bond_cash_flows(
             dt_market_close_date, 
             self.balance, 
             self.coupon, 
             self.maturity_years
         )
 
-        str_payment_dates, str_cash_flows = create_spot_semi_bond_cash_flows(
+        str_payment_dates, str_cash_flows = create_semi_bond_cash_flows(
             str_market_close_date, 
             self.balance, 
             self.coupon, 
@@ -142,10 +141,10 @@ class TestCreateSpotSemiBondCashFlows(unittest.TestCase):
         final payment date and cash flow are accurate.
         """
         short_maturity_years = 0.5
-        expected_dates = [np.datetime64('2025-02-09', 'D')]
+        expected_dates = [np.datetime64('2025-02-10', 'D')]
         expected_cash_flows = np.array([1025000.0])
         
-        payment_dates, cash_flows = create_spot_semi_bond_cash_flows(
+        payment_dates, cash_flows = create_semi_bond_cash_flows(
             self.market_close_date, 
             self.balance, 
             self.coupon, 
@@ -153,58 +152,10 @@ class TestCreateSpotSemiBondCashFlows(unittest.TestCase):
         )
         
         # Check if the payment dates match
-        self.assertEqual(payment_dates, expected_dates)
+        np.testing.assert_array_equal(payment_dates, expected_dates)
         
         # Check if the cash flows match
         np.testing.assert_array_equal(cash_flows, expected_cash_flows)
-
-class TestDiscountCashFlows(unittest.TestCase):
-
-    def setUp(self):
-        # Set up common test data
-        self.payment_dates = [
-            np.datetime64('2025-02-13', 'D'),
-            np.datetime64('2025-08-13', 'D'),
-            np.datetime64('2026-02-13', 'D'),
-            np.datetime64('2026-08-13', 'D'),
-            np.datetime64('2027-02-13', 'D')
-        ]
-        self.cash_flows = np.array([10000, 10000, 10000, 10000, 110000])  # 10,000 every 6 months, final payment 110,000
-        self.discount_rate_vals = np.array([0.02, 0.025, 0.03, 0.035, 0.04])  # Discount rates corresponding to each payment date
-        self.discount_rate_dates = [
-            np.datetime64('2025-02-13', 'D'),
-            np.datetime64('2025-08-13', 'D'),
-            np.datetime64('2026-02-13', 'D'),
-            np.datetime64('2026-08-13', 'D'),
-            np.datetime64('2027-02-13', 'D')
-        ]
-
-    def test_basic_functionality(self):
-        """
-        Test that the discount_cash_flows function calculates the present value correctly.
-        """
-        expected_present_value = 143420.645  # Replace with expected present value calculated manually or with a trusted tool
-        present_value = discount_cash_flows(
-            self.payment_dates,
-            self.cash_flows,
-            self.discount_rate_vals,
-            self.discount_rate_dates
-        )
-        self.assertAlmostEqual(present_value, expected_present_value, places=2)
-
-    def test_mismatched_array_lengths(self):
-        """
-        Test that the function raises a ValueError when the length of cash flows
-        does not match the number of payment dates.
-        """
-        mismatched_cash_flows = np.array([10000, 10000, 10000, 10000])  # One less cash flow
-        with self.assertRaises(ValueError):
-            discount_cash_flows(
-                self.payment_dates,
-                mismatched_cash_flows,
-                self.discount_rate_vals,
-                self.discount_rate_dates
-            )
         
 if __name__ == "__main__":
     unittest.main()
